Perft Improvements

+++ src/perft.h
@@ -5,8 +5,8 @@
 #include "board.h"
 #include "movegen.h"
 
-int perft(Board& board, int depth);
-int perft_divide(Board& board, int depth);
+uint64_t perft(Board& board, int depth);
+uint64_t perft_divide(Board& board, int depth);
 void perft_debug(Board& board, int depth);
 
 #endif
